% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{coef,Ib-method}
\alias{coef,Ib-method}
\title{Method for extracting coefficients from an object in class union "Ib"}
\usage{
\S4method{coef}{Ib}(object, ...)
}
\arguments{
\item{object}{an object of class union "Ib"}

\item{...}{further arguments to pass to \code{coef}}
}
\description{
Method for extracting coefficients from an object in class union "Ib"
}
\seealso{
\linkS4class{Ib}, \link[stats]{coef}
}
