% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{effects,Ib-method}
\alias{effects,Ib-method}
\title{Method for extracting effects from an object in class union "Ib"}
\usage{
\S4method{effects}{Ib}(object, ...)
}
\arguments{
\item{object}{an object of class union "Ib"}

\item{...}{further arguments to pass to \code{effects}}
}
\description{
Method for extracting effects from an object in class union "Ib"
}
\seealso{
\linkS4class{Ib}, \link[stats]{effects}
}
