% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{fitted,Ib-method}
\alias{fitted,Ib-method}
\title{Method for extracting fitted values from an object in class union "Ib"}
\usage{
\S4method{fitted}{Ib}(object, ...)
}
\arguments{
\item{object}{an object of class union "Ib"}

\item{...}{further arguments to pass to \code{fitted}}
}
\description{
Method for extracting fitted values from an object in class union "Ib"
}
\seealso{
\linkS4class{Ib}, \link[stats]{fitted.values}
}
