% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{getEst}
\alias{getEst}
\alias{getEst,Ib-method}
\title{Accessor to the object in class union "Ib"}
\usage{
getEst(x)

\S4method{getEst}{Ib}(x)
}
\arguments{
\item{x}{an object of class union "Ib"}
}
\value{
an estimate (as in \code{\link{getExtra}}).
}
\description{
Method for obtaining estimates from fitted model within
any object of class union \linkS4class{Ib}.
}
\details{
This methods allow to access extra parameter
estimates. If \code{extra_param=TRUE}, it becomes equivalent
to \code{\link{coef}}.
}
\seealso{
\linkS4class{Ib}
}
