% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{getExtra}
\alias{getExtra}
\alias{getExtra,Ib-method}
\title{Accessor to an extra part in class union "Ib"}
\usage{
getExtra(x)

\S4method{getExtra}{Ib}(x)
}
\arguments{
\item{x}{an object of class union "Ib"}
}
\value{
a \code{list} with the following components:
\tabular{ll}{
iteration \tab number of iterations (\eqn{k})\cr
of \tab value of the objective function
\eqn{||\hat{\pi}-\frac{1}{H}\sum_{h=1}^H\hat{\pi}_h(\hat{\theta}^{k})||}{%
||\pi-\frac{1}{H}\sum_{h=1}^H\pi_h(\theta^{k})||}\cr
estimate\tab value of the estimates \eqn{\hat{\theta}^{k}}{\theta^{k}}\cr
test_theta\tab value for difference of thetas:
\eqn{||\hat{\theta}^k-\hat{\theta}^{k-1}||}{||\theta^k-\theta^{k-1}||}\cr
ib_warn\tab optional warning message\cr
boot\tab \code{matrix} of \eqn{H} bootstrap estimates:
\eqn{\hat{\pi}(\hat{\theta}^k)}{\pi(\theta^k)}}
}
\description{
Method for obtaining a extra values generated by
the iterative bootstrap procedure within
any object of class union \linkS4class{Ib}.
}
\seealso{
\linkS4class{Ib}
}
