% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{getObject}
\alias{getObject}
\alias{getObject,Ib-method}
\title{Accessor to the object in class union "Ib"}
\usage{
getObject(x)

\S4method{getObject}{Ib}(x)
}
\arguments{
\item{x}{an object of class union "Ib"}
}
\description{
Method for obtaining a fitted model within
any object of class union \linkS4class{Ib}.
}
\seealso{
\linkS4class{Ib}
}
