% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ib_glm.R
\name{ib,negbin-method}
\alias{ib,negbin-method}
\title{\code{\link{ib}} method for \code{negbin} object
from \code{\link[MASS]{glm.nb}} function of \pkg{MASS}
package.}
\usage{
\S4method{ib}{negbin}(object, thetastart = NULL, control = list(...), extra_param = FALSE, ...)
}
\arguments{
\item{object}{an \code{object} representing a fitted model (see 'Details').}

\item{thetastart}{an optional starting value for the iterative procedure.
If \code{NULL} (default), the procedure starts at the estimates in \code{object}.}

\item{control}{a \code{list} of parameters for controlling the iterative procedure
(see \code{\link{ibControl}}).}

\item{extra_param}{if \code{TRUE}, the bias of estimation of extra parameters
is performed (see 'Details').}

\item{...}{additional optional arguments (see 'Details').}
}
\description{
\code{\link{ib}} method for \code{negbin} object
from \code{\link[MASS]{glm.nb}} function of \pkg{MASS}
package.
}
