% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ib.R
\name{ib}
\alias{ib}
\title{Bias correction via iterative bootstrap}
\usage{
ib(object, thetastart = NULL, control = list(...), ...)
}
\arguments{
\item{object}{an object representing a fitted model.}

\item{...}{additional optional arguments.}
}
\value{
The \code{object} with bias corrected estimators.
}
\description{
Correct the bias of a fitted model object with the iterative
bootstrap procedure.
}
