% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ib.R
\name{ibControl}
\alias{ibControl}
\title{Auxiliary for controlling IB}
\usage{
ibControl(
  tol = 1e-05,
  maxit = 25,
  verbose = FALSE,
  cens = FALSE,
  right = NULL,
  left = NULL,
  seed = 123L,
  H = 1L
)
}
\description{
Auxiliary for controlling IB
}
