% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{IbVglm-class}
\alias{IbVglm-class}
\title{An S4 class to represent a vglm object to which ib is applied}
\description{
An S4 class to represent a vglm object to which ib is applied
}
\section{Slots}{

\describe{
\item{\code{object}}{a \link[VGAM]{vglm} object with bias corrected quantities}

\item{\code{ib_extra}}{a \code{list} with additional information from the
iterative bootstrap procedure}
}}

\author{
Samuel Orso
}
