% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{plot,Ib,ANY-method}
\alias{plot,Ib,ANY-method}
\title{Method for plotting an object in class union "Ib"}
\usage{
\S4method{plot}{Ib,ANY}(x, y = NULL, ...)
}
\arguments{
\item{x}{an object of class union "Ib"}

\item{y}{not used}

\item{...}{further arguments to pass to \code{plot}}
}
\description{
Method for plotting an object in class union "Ib"
}
\seealso{
\linkS4class{Ib}, \link[stats]{plot.lm}
}
