% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{predict,Ib-method}
\alias{predict,Ib-method}
\title{Method for making predictions from an object in class union "Ib"}
\usage{
\S4method{predict}{Ib}(object, ...)
}
\arguments{
\item{object}{an object of class union "Ib"}

\item{...}{further arguments to pass to \code{predict}}
}
\description{
Method for making predictions from an object in class union "Ib"
}
\seealso{
\linkS4class{Ib}, \link[stats]{predict}
}
