% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{residuals,Ib-method}
\alias{residuals,Ib-method}
\title{Method for extracting residuals from an object in class union "Ib"}
\usage{
\S4method{residuals}{Ib}(object, ...)
}
\arguments{
\item{object}{an object of class union "Ib"}

\item{...}{further arguments to pass to \code{residuals}}
}
\description{
Method for extracting residuals from an object in class union "Ib"
}
\seealso{
\linkS4class{Ib}, \link[stats]{residuals}
}
