% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_defaults_ib.R
\name{scale_defaults_ib}
\alias{scale_defaults_ib}
\alias{scale_x_c_ib}
\alias{scale_x_d_ib}
\alias{scale_y_c_ib}
\alias{scale_y_d_ib}
\title{Custom scale default settings}
\usage{
scale_x_c_ib(...)

scale_x_d_ib(...)

scale_y_c_ib(...)

scale_y_d_ib(...)
}
\description{
Sets default values for the \code{expand} argument of scale_(x|y)\emph{continuous and
scale}(x|y)_discrete.
}
