% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{show,Ib-method}
\alias{show,Ib-method}
\title{Method for printing object in class union "Ib"}
\usage{
\S4method{show}{Ib}(object)
}
\arguments{
\item{object}{an object of class union "Ib"}
}
\description{
Method for printing object in class union "Ib"
}
\seealso{
\linkS4class{Ib}
}
