% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{show,SummaryIb-method}
\alias{show,SummaryIb-method}
\title{Summarizing a fitted model corrected by the ib procedure}
\usage{
\S4method{show}{SummaryIb}(object)
}
\arguments{
\item{object}{a summary object of member of \linkS4class{SummaryIb}}
}
\description{
Method for printing a \code{summary} of
class union \linkS4class{SummaryIb}.
}
\seealso{
\linkS4class{SummaryIb}
}
