% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ib_glmer.R
\name{simulation,glmerMod-method}
\alias{simulation,glmerMod-method}
\title{Simulation for linear mixed model regression}
\usage{
\S4method{simulation}{glmerMod}(object, control = list(...), ...)
}
\arguments{
\item{object}{an object of class \code{IbGlmer}, see \linkS4class{Ib}}

\item{control}{a \code{list} of parameters for controlling the iterative procedure
(see \code{\link{ibControl}}).}

\item{...}{further arguments.}
}
\description{
simulation method for class \code{IbGlmer}, see \linkS4class{Ib}
}
