% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ib_nls.R
\name{simulation,nls-method}
\alias{simulation,nls-method}
\title{Simulation for nonlinear regression}
\usage{
\S4method{simulation}{nls}(object, control = list(...), std = NULL, ...)
}
\arguments{
\item{object}{an object of class \linkS4class{IbNls}}

\item{control}{a \code{list} of parameters for controlling the iterative procedure
(see \code{\link{ibControl}}).}

\item{std}{\code{NULL} by default; standard deviation to pass to simulation.}

\item{...}{further arguments}
}
\description{
simulation method for class \linkS4class{IbNls}
}
