% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ib_vglm.R
\name{simulation,vglm-method}
\alias{simulation,vglm-method}
\title{Simulation for vector generalized linear model regression}
\usage{
\S4method{simulation}{vglm}(object, control = list(...), extra_param = NULL, ...)
}
\arguments{
\item{object}{an object of class \linkS4class{IbVglm}}

\item{control}{a \code{list} of parameters for controlling the iterative procedure
(see \code{\link{ibControl}}).}

\item{extra_param}{\code{NULL} by default; extra parameters to pass to simulation.}

\item{...}{further arguments}
}
\description{
simulation method for class \linkS4class{IbVglm}
}
