% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary,IbBetareg-method}
\alias{summary,IbBetareg-method}
\title{Summarizing a beta regression fit corrected by
the iterative bootstrap}
\usage{
\S4method{summary}{IbBetareg}(object, ...)
}
\arguments{
\item{object}{an object of class \linkS4class{IbBetareg}}

\item{...}{further arguments passed to \code{summary.betareg}}
}
\description{
summary method for class \linkS4class{IbBetareg}
}
\seealso{
\link[betareg]{summary.betareg}
}
