% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary,IbGlmer-method}
\alias{summary,IbGlmer-method}
\title{Summarizing a generalized linear mixed model regression fit corrected by
the iterative bootstrap}
\usage{
\S4method{summary}{IbGlmer}(object, ...)
}
\arguments{
\item{object}{an object of class \linkS4class{IbGlmer}}

\item{...}{further arguments passed to \code{summary.merMod} of \pkg{lme4}}
}
\description{
summary method for class \linkS4class{IbGlmer}
}
