% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary,IbLmer-method}
\alias{summary,IbLmer-method}
\title{Summarizing a linear mixed model regression fit corrected by
the iterative bootstrap}
\usage{
\S4method{summary}{IbLmer}(object, ...)
}
\arguments{
\item{object}{an object of class \linkS4class{IbLmer}}

\item{...}{further arguments passed to \code{summary.merMod} of \pkg{lme4}}
}
\description{
summary method for class \linkS4class{IbLmer}
}
