% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary,IbNegbin-method}
\alias{summary,IbNegbin-method}
\title{Summarizing a negative binomial regression fits corrected by
the iterative bootstrap}
\usage{
\S4method{summary}{IbNegbin}(object, ...)
}
\arguments{
\item{object}{an object of class \linkS4class{IbNegbin}}

\item{...}{further arguments passed to \code{summary.negbin}}
}
\description{
summary method for class \linkS4class{IbNegbin}
}
\seealso{
\link[MASS]{summary.negbin}
}
