% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary,IbNls-method}
\alias{summary,IbNls-method}
\title{Summarizing a nonlinear regression fit corrected by
the iterative bootstrap}
\usage{
\S4method{summary}{IbNls}(object, ...)
}
\arguments{
\item{object}{an object of class \linkS4class{IbNls}}

\item{...}{further arguments passed to \code{summary.nls} of \pkg{stats}}
}
\description{
summary method for class \linkS4class{IbNls}
}
