% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SummaryIb-class}
\alias{SummaryIb-class}
\alias{SummaryIbBetareg-class}
\alias{SummaryIbGlm-class}
\alias{SummaryIbLm-class}
\alias{SummaryIbLmer-class}
\alias{SummaryIbGlmer-class}
\alias{SummaryIbNegbin-class}
\alias{SummaryIbNls-class}
\alias{SummaryIbVglm-class}
\alias{SummaryIb}
\title{Summaries for \code{ib} wrappers}
\description{
S4 classes that wrap summary objects corresponding to the \code{Ib*} classes.

Members: \linkS4class{SummaryIbBetareg}, \linkS4class{SummaryIbGlm},
\linkS4class{SummaryIbLm}, \linkS4class{SummaryIbLmer},
\linkS4class{SummaryIbGlmer}, \linkS4class{SummaryIbNegbin},
\linkS4class{SummaryIbNls}, \linkS4class{SummaryIbVglm}.
}
\section{Slots}{

\describe{
\item{\code{summ}}{The \code{summary.*} object.}

\item{\code{ib_extra}}{A \code{list} of extra metadata from \code{ib}.}
}}

\seealso{
\linkS4class{SummaryIb}, \linkS4class{Ib}
}
