% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ib.R
\name{theme_ib}
\alias{theme_ib}
\title{A \link{ggplot2} theme for Isabella Benabaye's personal use}
\usage{
theme_ib(
  title_family = "IBM Plex Sans",
  text_family = "IBM Plex Mono",
  base_size = 13,
  text_color = "gray20",
  bg_color = "#F7F7F7",
  line_color = "gray50",
  plot_margin = margin(20, 20, 20, 20),
  plots_pane = FALSE,
  md = FALSE
)
}
\arguments{
\item{title_family}{title elements font family}

\item{text_family}{text elements font family}

\item{base_size}{base font size}

\item{text_color}{color of the text & title elements}

\item{bg_color}{plot background color}

\item{line_color}{color of line elements}

\item{plot_margin}{plot margin specifications}

\item{plots_pane}{indicate whether plots will be viewed in plots pane or not}

\item{md}{indicate whether to use markdown elements for text}
}
\description{
\code{theme_ib} is a \link{ggplot2} theme in Isabella Benabaye's personal style.
}
\details{
By default (\code{plots_pane = FALSE}), the theme adjusts the text sizes for
printing.\code{plots_pane = TRUE} is meant to be used when viewing plots in the
plots pane and text sizes are not adjusted.  There is also an option (\code{md = TRUE}) to use markdown theme elements from \code{ggtext} instead of
\code{element_text()}.
}
