% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_geom_fonts_ib.R
\name{update_geom_fonts_ib}
\alias{update_geom_fonts_ib}
\title{Update font defaults for text geoms to match theme_ib}
\usage{
update_geom_fonts_ib(
  family = "IBM Plex Sans",
  face = "plain",
  size = 7,
  color = "gray20",
  hjust = 0
)
}
\arguments{
\item{family}{default geom font family}

\item{face}{default geom font face}

\item{size}{default geom font size}

\item{color}{default text color}

\item{hjust}{default hjust value}
}
\description{
Updates \link[ggplot2:geom_text]{ggplot2::geom_label}, \link[ggplot2:geom_text]{ggplot2::geom_text}, and \link[ggplot2:annotate]{ggplot2::annotate}
font defaults
}
