% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{vcov,Ib-method}
\alias{vcov,Ib-method}
\title{Method for calculating covariance matrix from an object in class union "Ib"}
\usage{
\S4method{vcov}{Ib}(object, ...)
}
\arguments{
\item{object}{an object of class union "Ib"}

\item{...}{further arguments to pass to \code{vcov}}
}
\description{
Method for calculating covariance matrix from an object in class union "Ib"
}
\seealso{
\linkS4class{Ib}, \link[stats]{vcov}
}
