% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Parametric bootstrap}
\usage{
bootstrap(object, B = 1000, extra_param = FALSE, ...)
}
\arguments{
\item{object}{an \code{object} representing a fitted model (see 'Details').}

\item{B}{an \code{integer} for number of bootstrap replicates (default
1,000).}

\item{extra_param}{if \code{TRUE}, bootstrap is also performed for extra parameters
(see 'Details').}

\item{...}{additional optional arguments to pass to \code{ibControl}.}
}
\value{
A \code{matrix} p (size of parameter) times B of bootstrapped estimates.
}
\description{
Method for generating parametric bootstrap estimates from
a fitted model.
}
\details{
This method is a simple wrapper around the \code{ib} method
where number of iterations is set to 1.
}
\seealso{
\code{\link{ib}}, \code{\link{ibControl}}
}
\author{
Samuel Orso
}
