% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{getEst}
\alias{getEst}
\alias{getEst,Ib-method}
\title{Accessor to the object in class union "Ib"}
\usage{
getEst(x)

\S4method{getEst}{Ib}(x)
}
\arguments{
\item{x}{an object of class union "Ib"}
}
\description{
Method for obtaining estimates from fitted model within
any object of class union \linkS4class{Ib}.
}
\seealso{
\linkS4class{Ib}
}
