% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{getExtra}
\alias{getExtra}
\alias{getExtra,Ib-method}
\title{Accessor to an extra part in class union "Ib"}
\usage{
getExtra(x)

\S4method{getExtra}{Ib}(x)
}
\arguments{
\item{x}{an object of class union "Ib"}
}
\description{
Method for obtaining a extra values generated by
the iterative bootstrap procedure within
any object of class union \linkS4class{Ib}.
}
\seealso{
\linkS4class{Ib}
}
