% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{IbGlm-class}
\alias{IbGlm-class}
\alias{IbLm-class}
\alias{IbLmer-class}
\alias{IbNegbin-class}
\alias{IbNls-class}
\alias{IbVglm-class}
\alias{Ib-class}
\title{An S4 class union for \code{ib}}
\value{
Each member of the union has a \code{slot} with the initial object
corrected by the \code{ib} (see \code{\link{getObject}}) and a second \code{slot} with
extra meta data from \code{ib} (see \code{\link{getExtra}}).
}
\description{
Members of the union are \linkS4class{IbGlm}, \linkS4class{IbLm},
\linkS4class{IbLmer}, \linkS4class{IbNegbin}, \linkS4class{IbNls},
\linkS4class{IbVglm}
}
\details{
The `Functions` section describes members of the class union.
}
\section{Functions}{
\itemize{
\item \code{IbGlm-class}: fitted model by \code{glm} from \pkg{stats}

\item \code{IbLm-class}: fitted model by \code{lm} from \pkg{stats}

\item \code{IbLmer-class}: fitted model by \code{lmer} from \pkg{lme4}

\item \code{IbNegbin-class}: fitted model by \code{glm.nb} from \pkg{MASS}

\item \code{IbNls-class}: fitted model by \code{nls} from \pkg{stats}

\item \code{IbVglm-class}: fitted model by \code{vglm} from \pkg{VGAM}
}}

\seealso{
\code{\link{getExtra}}, \code{\link{getObject}}
}
\author{
Samuel Orso
}
