% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{Ib-class}
\alias{Ib-class}
\title{An S4 class union}
\description{
Members of the union are \linkS4class{IbGlm}, \linkS4class{IbLm},
\linkS4class{IbLmer}, \linkS4class{IbNegbin}, \linkS4class{IbVglm}
}
