% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{Ib-class}
\alias{Ib-class}
\title{An S4 class union for \code{ib}}
\value{
Each member of the union has a \code{slot} with the initial object
corrected by the \code{ib} (see \code{\link{getObject}}) and a second \code{slot} with
extra meta data from \code{ib} (see \code{\link{getExtra}}).
}
\description{
Members of the union are \code{IbBetareg}, \code{IbGlm},
\code{IbLm}, \code{IbLmer}, \code{IbNegbin},
\code{IbNls}, \code{IbVglm}
}
\details{
The `Functions` section describes members of the class union.
}
\seealso{
\code{\link{getExtra}}, \code{\link{getObject}}
}
\author{
Samuel Orso
}
