% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ib.R
\name{ib}
\alias{ib}
\title{Bias  correction via the iterative bootstrap algorithm}
\usage{
ib(
  object,
  ibcontrol = ib_control(),
  method = "IB",
  theta_start = NULL,
  simcontrol = sim_control(),
  ...
)
}
\description{
Bias  correction via the iterative bootstrap algorithm
}
\author{
Samuel Orso
}
