% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ib.R, R/ib_glm.R, R/ib_lm.R, R/ib_lmer.R
\name{ib}
\alias{ib}
\alias{ib.glm}
\alias{ib.lm}
\alias{ib.lmerMod}
\title{Bias correction via iterative bootstrap}
\usage{
ib(object, thetastart = NULL, control = list(...), ...)

\method{ib}{glm}(
  object,
  thetastart = NULL,
  control = list(...),
  shape = FALSE,
  overdispersion = FALSE,
  ...
)

\method{ib}{lm}(object, thetastart = NULL, control = list(...), var = FALSE, ...)

\method{ib}{lmerMod}(object, thetastart = NULL, control = list(...), Sigma = FALSE, ...)
}
\arguments{
\item{object}{an \code{object} representing a fitted model (see 'Details').}

\item{thetastart}{an optional starting value for the iterative procedure.
If \code{NULL} (default), the procedure starts at the estimates in \code{object}.}

\item{control}{a list of parameters for controlling the iterative procedure
(see \code{\link{ibControl}}).}

\item{...}{additional optional arguments (see 'Details').}

\item{shape}{if \code{TRUE}, the shape parameter for the \code{\link[stats:family]{Gamma}}
family is also corrected}

\item{overdispersion}{if \code{TRUE}, the overdispersion parameter of the
negative binomial regression in \code{\link[MASS]{glm.nb}} is also corrected}

\item{var}{if \code{TRUE}, the variance of the residuals in \link[stats]{lm} is
also corrected}

\item{Sigma}{if \code{TRUE}, all the random effects (variances and correlations) and the variance
of the residuals in \link[lme4]{lmer} are also corrected}
}
\value{
A fitted model \code{object} where estimates are bias corrected with the \code{ib}.
}
\description{
\code{ib} is used to correct the bias of a fitted model \code{object}
with the iterative bootstrap procedure.
}
\details{
The iterative bootstrap procedure is described in
\insertCite{kuk1995;textual}{ib} and further
studied by \insertCite{guerrier2019;textual}{ib} and
\insertCite{guerrier2020;textual}{ib}. The \emph{k}th iteration of this
algorithm is
\deqn{\hat{\theta}^{k} = \hat{\theta}^{k-1} + \hat{\pi} -
\frac{1}{H}\sum_{h=1}^H\hat{\pi}_h(\hat{\theta}^{k-1})}{%
\theta^(k)=\theta^(k-1) + \pi - \sum \pi_h(\theta^(k-1)) / H
}
for \eqn{k=1,2,\ldots} and where the sum is over \eqn{h=1,\ldots,H}.
The estimate \eqn{\pi} is provided by the \code{object}.
The value \eqn{\pi_h(\theta)} is a parametric bootstrap
estimate where the bootstrap sample is generated from \eqn{\theta}
and a fixed \code{seed} (see \code{\link{ibControl}}).
The greater the parameter value \eqn{H} generally the better bias correction
but the more computation it requires (see \code{\link{ibControl}}).
If \code{thetastart=NULL}, the initial value of the procedure is \eqn{\theta^(0)=\pi}.
The number of iterations are controlled by \code{maxit} parameter of \code{\link{ibControl}}.

By default, the method correct \code{\link[stats:coef]{coefficients}} only. For
extra parameters, it depends on the model. Currently, \code{ib}
supports the following \code{object}:
\tabular{ll}{
   \code{\link[stats]{glm}}\tab
       with \code{shape=TRUE}, the shape parameter for the \code{\link[stats:family]{Gamma}}
       family is also corrected. Note that the \code{\link[stats:family]{quasi}} families
       are not supported for the moment as they have no simulation method
       (see \code{\link[stats]{simulate}}).
   \cr
   \code{\link[MASS]{glm.nb}}\tab
       with \code{overdispersion=TRUE}, the overdispersion parameter of the
       negative binomial regression is also corrected.
   \cr
   \code{\link[stats]{lm}}\tab
       with \code{vars=TRUE}, the variance of the residuals is
       also corrected. Note that using the \code{ib} is not useful as coefficients
       are already unbiased, unless one considers different
       data generating mechanism such as censoring, missing values
       and outliers (see \code{\link{ibControl}}).
   \cr
   \code{\link[lme4]{lmer}}\tab
       by default, only the fixed effects are corrected. With \code{Sigma=TRUE},
       all the random effects (variances and correlations) and the variance
       of the residuals are also corrected. Note that using the \code{ib} is
       certainly not useful with the argument \code{REML=TRUE} in
       \code{\link[lme4]{lmer}} as the bias of variance components is
       already addressed, unless one considers different
       data generating mechanism such as censoring, missing values
       and outliers (see \code{\link{ibControl}}).
   \cr
}
}
\references{
\insertAllCited{}
}
