% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{simulation}
\alias{simulation}
\alias{simulation,Ib-method}
\title{Generic for simulating from the object}
\usage{
simulation(object, control = list(...), ...)

\S4method{simulation}{Ib}(object, control = list(...), ...)
}
\arguments{
\item{object}{an object of class union "Ib"}

\item{control}{a control list}

\item{...}{further argument to pass}
}
\value{
simulated responses.
}
\description{
Method for simulating responses from an object.
}
