% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SummaryIb-class}
\alias{SummaryIb-class}
\title{An S4 class union for \code{summary}}
\description{
Members of the union are \code{SummaryIbBetareg}, \code{SummaryIbGlm}, \code{SummaryIbLm},
\code{SummaryIbLmer}, \code{SummaryIbNegbin}, \code{SummaryIbNls},
\code{SummaryIbVglm}
}
\details{
The `Functions` section describes members of the class union.
}
\author{
Samuel Orso
}
