% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SummaryIbBetareg-class}
\alias{SummaryIbBetareg-class}
\alias{SummaryIbGlm-class}
\alias{SummaryIbLm-class}
\alias{SummaryIbLmer-class}
\alias{SummaryIbNegbin-class}
\alias{SummaryIbNls-class}
\alias{SummaryIbVglm-class}
\alias{SummaryIb-class}
\title{An S4 class union for \code{summary}}
\description{
Members of the union are \linkS4class{SummaryIbBetareg}, \linkS4class{SummaryIbGlm}, \linkS4class{SummaryIbLm},
\linkS4class{SummaryIbLmer}, \linkS4class{SummaryIbNegbin}, \linkS4class{SummaryIbNls},
\linkS4class{SummaryIbVglm}
iterative bootstrap procedure
}
\details{
The `Functions` section describes members of the class union.
}
\section{Functions}{
\itemize{
\item \code{SummaryIbBetareg-class}: summary of class \code{summary.betareg} from \pkg{betareg}

\item \code{SummaryIbGlm-class}: summary of class \code{summary.glm} from \pkg{stats}

\item \code{SummaryIbLm-class}: summary of class \code{summary.lm} from \pkg{stats}

\item \code{SummaryIbLmer-class}: summary of class \code{summary.merMod} from \pkg{lme4}

\item \code{SummaryIbNegbin-class}: summary of class \code{summary.negbin} from \pkg{MASS}

\item \code{SummaryIbNls-class}: summary of class \code{summary.nls} from \pkg{stats}

\item \code{SummaryIbVglm-class}: summary of class \code{summary.vglm} from \pkg{VGAM}
}}

\author{
Samuel Orso
}
